pipeline/requirements_clean.txt
numpy>=1.26.0
rasterio>=1.3.9
xarray>=2024.5.0
netCDF4>=1.6.5
shapely>=2.0.3
scikit-image>=0.24.0
requests>=2.32.3
richdem>=0.3.4
asf-search>=7.2.3
pillow>=10.3.0
